refactor(data/paths): split PathBuilder god-module into per-concern impl blocks
The unified `PathBuilder` facade in `data/paths/builder/mod.rs` was 920 LOC
with 89 `pub fn` delegators packed into a single file — ~9× the size of any
other Layer 3 builder. Audit (`audit-architecture` finding architecture-003)
flagged it as a god-module.

Split the delegators into a new `path_builder/` subdirectory, with one file
per concern (characteristic, filetype, security, filename, construction,
format, home, context, building, lenient, boundary), each containing an
`impl PathBuilder { ... }` extension block. This mirrors the split-impl
pattern already used in `observe/builder/` (7 files) and
`observe/context/builder/` (5 files).

The struct, constructors, and the five methods that emit metrics directly
(`detect`, `is_safe`, `validate_detailed`, `validate_path`, `sanitize`)
stay in `mod.rs` so the `define_metrics!`-generated `metric_names` module
remains in scope.

Public API is byte-identical: all 89 methods remain on `PathBuilder` via
merged impl blocks; no caller-site changes needed across the 151 use sites.
`mod.rs` drops from 920 → 386 LOC; each split file is well under 100 LOC.

Closes #174
